// shared/src/types.ts
export interface IProfile {
  _id: string;
  userId: string;
  fullName: string;
  tagline: string;
  bio: string;
  location: string;
  email: string;
  phone?: string;
  avatar?: string;
  resume?: string;
  socialLinks: ISocialLink[];
  visibility: boolean;
  createdAt: Date;
  updatedAt: Date;
  createdBy: string;
}

export interface ISocialLink {
  platform: string;
  url: string;
  icon?: string;
}